requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
websockets==15.0.1
opencv-python-headless==4.11.0.86
ultralytics==8.3.145
mediapipe==0.10.21
anthropic==0.52.0
numpy<2.1
Pillow==11.2.1
python-dotenv==1.1.0
pydantic==2.11.3
httpx==0.28.1
